+++ module/macros/_module.mjs
@@ -2,6 +2,7 @@ import { abilityGeneration } from './player/abilityGeneration.mjs';
 import { diceOracles } from './player/diceOracles.mjs'
 import { escalatorDie } from './gm/escalatorDie.mjs';
 import { generateRandomBackground } from './player/generateRandomBackground.mjs';
+import { characterGeneratorDialog } from './player/characterGenerator.mjs';
 import { giveCash } from './gm/giveCash.mjs';
 import { giveExperience } from './gm/giveExperience.mjs';
 import { giveHeroDice } from './gm/giveHeroDice.mjs';
@@ -20,7 +21,8 @@ const gm = {
 const player = {
   diceOracles,
   abilityGeneration,
-  generateRandomBackground
+  generateRandomBackground,
+  characterGeneratorDialog
 };
 
 export { gm, player };